CLN: Rename datastructure.export to datastructure._internal
