+++ src/fmu/dataio/datastructure/_internal/content.py
@@ -1,3 +1,11 @@
+"""
+This module, `datastructure._internal`, contains internal data structures that
+are designed to depend on external modules, but not the other way around.
+This design ensures modularity and flexibility, allowing external modules
+to be potentially separated into their own repositories without dependencies
+on the internals.
+"""
+
 from __future__ import annotations
 
 import warnings
